requirements_dev.txt
-i https://pypi.python.org/simple
apipkg==1.5
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
codecov==2.0.15
coverage==4.5.3
django-extensions==2.1.6
execnet==1.5.0
filelock==3.0.10
idna==2.8
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==6.0.0
multidict==4.5.2
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pydocstyle==3.0.0
pylint-celery==0.3
pylint-django==2.0.6
pylint-plugin-utils==0.5
pylint==2.3.1
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-forked==1.0.2
pytest-mock==1.10.1
pytest-vcr==1.0.1
pytest-xdist==1.26.1
pytest==4.0.2
python-slugify==1.2.6
pyyaml==3.13
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
toml==0.10.0
tox==3.7.0
transifex-client==0.13.6
typed-ast==1.3.1 ; implementation_name == 'cpython'
unidecode==1.0.23
urllib3==1.24.1
vcrpy==2.0.1
virtualenv==16.4.3
wrapt==1.11.1
yapf==0.26.0
yarl==1.3.0 ; python_version >= '3.4'
xmltodict
